+++ mocket/__init__.py
@@ -1,5 +1,6 @@
 from mocket.async_mocket import async_mocketize
-from mocket.mocket import FakeSSLContext, Mocket, MocketEntry, Mocketizer, mocketize
+from mocket.mocket import FakeSSLContext, Mocket, MocketEntry
+from mocket.mocketizer import Mocketizer, mocketize
 
 __all__ = (
     "async_mocketize",
